+++ guff/src/targets.rs
@@ -8,9 +8,8 @@ use dactyl::{
 	traits::BytesToUnsigned,
 };
 use oxford_join::OxfordJoinFmt;
-use lightningcss::targets::Browsers;
+use guff_css::lightningcss::targets::Browsers;
 use std::{
-	borrow::Borrow,
 	collections::HashMap,
 	fmt,
 	hash::{
@@ -51,20 +50,7 @@ include!(concat!(env!("OUT_DIR"), "/guff-browsers.rs"));
 /// * opera
 /// * safari
 /// * samsung (Samsung's Android browser)
-///
-/// ## Examples
-///
-/// ```
-/// use guff_css::Agents;
-///
-/// let agents = Agents::try_from("firefox 100, ie 11").unwrap();
-/// assert_eq!(agents.len(), 2);
-/// assert_eq!(agents.to_string(), "Firefox (100) and IE (11)");
-///
-/// // Invalid browser strings trigger an error.
-/// assert!(Agents::try_from("foobar 11").is_err());
-/// ```
-pub struct Agents(HashMap<Agent, (u32, u32), NoHash>);
+pub(super) struct Agents(HashMap<Agent, (u32, u32), NoHash>);
 
 impl fmt::Display for Agents {
 	fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
@@ -118,19 +104,16 @@ impl TryFrom<&str> for Agents {
 impl Agents {
 	#[must_use]
 	/// # Is Empty?
-	pub fn is_empty(&self) -> bool { self.0.is_empty() }
-
-	#[must_use]
-	/// # Length.
-	pub fn len(&self) -> usize { self.0.len() }
+	pub(super) fn is_empty(&self) -> bool { self.0.is_empty() }
 }
 
 impl Agents {
+	#[cfg(test)]
 	#[must_use]
 	/// # Get.
 	///
 	/// Return the major version the [`Agent`] is capped to, if any.
-	pub fn get(&self, agent: Agent) -> Option<u32> {
+	fn get(&self, agent: Agent) -> Option<u32> {
 		self.0.get(&agent).map(|(_, m)| *m)
 	}
 
@@ -138,26 +121,7 @@ impl Agents {
 	///
 	/// Restrict compatibility to a specific major browser release. If the
 	/// version is invalid or unknown, restrictions will be removed.
-	///
-	/// ## Examples
-	///
-	/// ```
-	/// use guff_css::{Agent, Agents};
-	///
-	/// let mut agents = Agents::default();
-	/// assert!(agents.is_empty()); // It starts with no restrictions.
-	///
-	/// // Cap Firefox to version 100.
-	/// agents.set(Agent::Firefox, 100);
-	/// assert_eq!(agents.len(), 1);
-	/// assert_eq!(Some(100), agents.get(Agent::Firefox));
-	///
-	/// // An invalid version (or zero) removes the restriction.
-	/// agents.set(Agent::Firefox, 0);
-	/// assert!(agents.is_empty());
-	/// assert!(agents.get(Agent::Firefox).is_none());
-	/// ```
-	pub fn set(&mut self, agent: Agent, major: u32) {
+	fn set(&mut self, agent: Agent, major: u32) {
 		if
 			0 < major &&
 			let Some(x) = agent.set().iter().rfind(|(_, m)| *m == major).copied()
@@ -165,50 +129,6 @@ impl Agents {
 			*(self.0.entry(agent).or_insert((0,0))) = x;
 		}
 		else { self.0.remove(&agent); }
-	}
-
-	/// # Cap Support to N Versions Back.
-	///
-	/// This method will restrict compatibility for the specified [`Agent`] to
-	/// its _nth_ release (working backwards).
-	///
-	/// For example, if the latest Chrome release is 101 and you pass a value
-	/// of one, compatibility will be capped to version 100. A value of two,
-	/// on the other hand, would cap to version 99, and so on.
-	///
-	/// If `n` is less than one, restrictions will be removed. If it exceeds
-	/// the number of stored releases — which are capped to a maximum of 16 —
-	/// the oldest known release will be used instead.
-	///
-	/// Because release data is baked in at compile time, this method may set
-	/// the target farther back in time than necessary, particularly for
-	/// browsers with fast release schedules, like Chrome and Firefox.
-	///
-	/// ## Examples
-	///
-	/// ```
-	/// use guff_css::{Agent, Agents};
-	///
-	/// let mut agents = Agents::default();
-	/// assert!(agents.is_empty()); // It starts with no restrictions.
-	///
-	/// // Cap Firefox to its penultimate version.
-	/// agents.set_nth(Agent::Firefox, 1);
-	/// assert_eq!(agents.len(), 1);
-	/// assert_eq!(Some(Agent::Firefox.nth(1)), agents.get(Agent::Firefox));
-	/// ```
-	pub fn set_nth(&mut self, agent: Agent, n: usize) {
-		if 0 < n {
-			let set = agent.set();
-			let v =
-				if n < set.len() { set[n] }
-				else { set[set.len() - 1] };
-
-			*(self.0.entry(agent).or_insert((0, 0))) = v;
-		}
-		else {
-			self.0.remove(&agent);
-		}
 	}
 }
 
@@ -239,9 +159,9 @@ impl From<Agents> for Option<Browsers> {
 
 
 #[repr(u8)]
-#[derive(Debug, Clone, Copy)]
+#[derive(Debug, Clone, Copy, Eq, PartialEq)]
 /// # Agent (Browser).
-pub enum Agent {
+pub(super) enum Agent {
 	/// # Generic Android Browser.
 	Android = 0_u8,
 
@@ -270,16 +190,6 @@ pub enum Agent {
 	Samsung,
 }
 
-impl AsRef<str> for Agent {
-	#[inline]
-	fn as_ref(&self) -> &str { self.as_str() }
-}
-
-impl Borrow<str> for Agent {
-	#[inline]
-	fn borrow(&self) -> &str { self.as_str() }
-}
-
 impl fmt::Display for Agent {
 	#[inline]
 	fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
@@ -287,16 +197,9 @@ impl fmt::Display for Agent {
 	}
 }
 
-impl Eq for Agent {}
-
 impl Hash for Agent {
 	#[inline]
 	fn hash<H: Hasher>(&self, state: &mut H) { state.write_u8(*self as u8); }
-}
-
-impl PartialEq for Agent {
-	#[inline]
-	fn eq(&self, other: &Self) -> bool { (*self as u8) == (*other as u8) }
 }
 
 impl TryFrom<&str> for Agent {
@@ -325,7 +228,7 @@ impl TryFrom<&str> for Agent {
 impl Agent {
 	#[must_use]
 	/// # As Str.
-	pub const fn as_str(self) -> &'static str {
+	pub(super) const fn as_str(self) -> &'static str {
 		match self {
 			Self::Android => "Android",
 			Self::Chrome => "Chrome",
@@ -337,28 +240,6 @@ impl Agent {
 			Self::Safari => "Safari",
 			Self::Samsung => "Samsung",
 		}
-	}
-
-	#[inline]
-	#[must_use]
-	/// # Latest Version.
-	///
-	/// Return the latest (known) major version for the browser.
-	pub const fn latest(self) -> u32 { self.nth(0) }
-
-	#[must_use]
-	/// # Nth.
-	///
-	/// Return the _nth_ (known) version of the browser, working backwards from
-	/// the latest (0).
-	///
-	/// To keep the code size down, a maximum of 16 releases are tracked. If `n`
-	/// is greater than the number of stored releases, the oldest stored will
-	/// be returned.
-	pub const fn nth(self, n: usize) -> u32 {
-		let set = self.set();
-		if n < set.len() { set[n].1 }
-		else { set[set.len() - 1].1 }
 	}
 
 	/// # Release Dataset.
@@ -374,5 +255,38 @@ impl Agent {
 			Self::Safari => &SAFARI,
 			Self::Samsung => &SAMSUNG,
 		}
+	}
+}
+
+
+
+#[cfg(test)]
+mod tests {
+	use super::*;
+
+	#[test]
+	fn t_agents() {
+		let agents = Agents::try_from("firefox 100, ie 11").unwrap();
+		assert_eq!(agents.0.len(), 2);
+		assert_eq!(agents.to_string(), "Firefox (100) and IE (11)");
+
+		// Invalid browser strings trigger an error.
+		assert!(Agents::try_from("foobar 11").is_err());
+	}
+
+	#[test]
+	fn t_agent() {
+		let mut agents = Agents::default();
+		assert!(agents.is_empty()); // It starts with no restrictions.
+
+		// Cap Firefox to version 100.
+		agents.set(Agent::Firefox, 100);
+		assert_eq!(agents.0.len(), 1);
+		assert_eq!(Some(100), agents.get(Agent::Firefox));
+
+		// An invalid version (or zero) removes the restriction.
+		agents.set(Agent::Firefox, 0);
+		assert!(agents.is_empty());
+		assert!(agents.get(Agent::Firefox).is_none());
 	}
 }